requirements.txt
azure-common==1.1.28
azure-core==1.24.1
azure-ai-formrecognizer==3.3.0b1
azure-identity==1.10.0
azure-keyvault==4.2.0
azure-keyvault-certificates==4.6.0
azure-keyvault-keys==4.7.0
azure-keyvault-secrets==4.6.0
azure-mgmt-core==1.3.1
azure-mgmt-datafactory==2.5.0
azure-storage-blob==12.13.0b1
azure-storage-file-datalake==12.5.0
azure-storage-file-share==12.6.0
Office365-REST-Python-Client==2.3.13
selenium==4.11.2
pandas==1.3.4
openpyxl==3.0.10
xlrd==2.0.1
requests==2.31.0
xlsxwriter==3.0.1
paramiko==3.4.0
zipfile39==0.0.8.0
